Create a subpackage for utils
